feat: 유저/파트너 사용완료쿠폰 api 분리
